Implemented the Virtual Ring Buffer, plus unit tests.
It isn't a FIFO implementation, with enqueuing and dequeuing. Its purpose is
to store the history of a stream so that any point a certain distance in the
past or later can be accessed. It still has what is an equivalent to enqueuing,
called advance. It also has a "past" method, for controlled access to the
stream's past.